+++ konsent/__main__.py
@@ -1,5 +1,1 @@
-from konsent import main
-
-
-if __name__ == '__main__':
-    main()
+from konsent.cli import main
